+++ src/zenml/integrations/kubeflow/orchestrators/kubeflow_orchestrator.py
@@ -35,11 +35,14 @@ from zenml.io import fileio
 from zenml.logger import get_logger
 from zenml.orchestrators import BaseOrchestrator
 from zenml.orchestrators.utils import create_tfx_pipeline
+from zenml.utils import networking_utils
 
 if TYPE_CHECKING:
     from zenml.pipelines.base_pipeline import BasePipeline
 
 logger = get_logger(__name__)
+
+DEFAULT_KFP_UI_PORT = 8080
 
 
 @orchestrator_store_factory.register(OrchestratorTypes.kubeflow)
@@ -47,7 +50,7 @@ class KubeflowOrchestrator(BaseOrchestrator):
     """Orchestrator responsible for running pipelines using Kubeflow."""
 
     custom_docker_base_image_name: Optional[str] = None
-    kubeflow_pipelines_ui_port: int = 8080
+    kubeflow_pipelines_ui_port: int = DEFAULT_KFP_UI_PORT
     kubernetes_context: Optional[str] = None
 
     def get_docker_image_name(self, pipeline_name: str) -> str:
@@ -292,10 +295,19 @@ class KubeflowOrchestrator(BaseOrchestrator):
         local_deployment_utils.deploy_kubeflow_pipelines(
             kubernetes_context=kubernetes_context
         )
+
+        port = self.kubeflow_pipelines_ui_port
+        if port == DEFAULT_KFP_UI_PORT and not networking_utils.port_available(
+            port
+        ):
+            # if the user didn't specify a specific port and the default port
+            # is occupied, fallback to a random open port
+            port = networking_utils.find_available_port()
+
         local_deployment_utils.start_kfp_ui_daemon(
             pid_file_path=self._pid_file_path,
             log_file_path=self._daemon_log_file_path,
-            port=self.kubeflow_pipelines_ui_port,
+            port=port,
         )
 
     def down(self) -> None:
