+++ src/zenml/utils/networking_utils.py
@@ -1,0 +1,39 @@
+#  Copyright (c) ZenML GmbH 2022. All Rights Reserved.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at:
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+#  or implied. See the License for the specific language governing
+#  permissions and limitations under the License.
+import socket
+
+from zenml.logger import get_logger
+
+logger = get_logger(__name__)
+
+
+def port_available(port: int) -> bool:
+    """Checks if a local port is available."""
+    try:
+        with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+            s.bind(("127.0.0.1", port))
+    except socket.error as e:
+        logger.debug("Port %d unavailable: %s", port, e)
+        return False
+
+    return True
+
+
+def find_available_port() -> int:
+    """Finds a local unoccupied port."""
+    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
+        s.bind(("127.0.0.1", 0))
+        _, port = s.getsockname()
+
+    return port
